- Cleared up a bunch of x64 warnings. - Ugh disabling the annoying yaml-cpp warning until it's fixed.

+++ src/Basescape/CraftArmorState.cpp
@@ -43,7 +43,7 @@ namespace OpenXcom
  * @param base Pointer to the base to get info from.
  * @param craft ID of the selected craft.
  */
-CraftArmorState::CraftArmorState(Game *game, Base *base, unsigned int craft) : State(game), _base(base), _craft(craft)
+CraftArmorState::CraftArmorState(Game *game, Base *base, size_t craft) : State(game), _base(base), _craft(craft)
 {
 	// Create objects
 	_window = new Window(this, 320, 200, 0, 0);
